+++ Classes/Class_Mage.lua
@@ -31,6 +31,6 @@ end
 -- Assignment model: Mage duties. Wids are stable.
 if AegisRP.Assign then
     local D = AegisRP.Assign.RegisterDuty
-    D{ key="INTELLECT", wid=4,  class="MAGE", tab="raidbuff", spell="Arcane Intellect", target="none", multi=false, dur=30*60 }
-    D{ key="SCORCH",    wid=13, class="MAGE", tab="debuff",   spell="Scorch",           target="none", multi=true,  dur=30 }
+    D{ key="INTELLECT", wid=4,  class="MAGE", tab="raidbuff", spell="Arcane Intellect", target="none", multi=false, dur=30*60, icon="Interface\\Icons\\Spell_Holy_MagicalSentry" }
+    D{ key="SCORCH",    wid=13, class="MAGE", tab="debuff",   spell="Scorch",           target="none", multi=true,  dur=30, icon="Interface\\Icons\\Spell_Fire_SoulBurn" }
 end